backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==4.2.20
django-cors-headers==4.7.0
djangorestframework==3.16.0
dnspython==2.7.0
drf-yasg==1.21.10
exceptiongroup==1.2.2
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
pymongo==4.11.3
pytest==8.3.5
pytest-django==4.10.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
